+++ plugin-packs/postcss-preset-env/src/lib/ids-by-execution-order.mjs
@@ -1,5 +1,7 @@
 // ids ordered by required execution, then alphabetically
 export default [
+	'mixins',
+
 	'custom-media-queries',
 	'environment-variables', // run environment-variables here to access transpiled custom media params and properties
 	'image-set-function', // run images-set-function before nesting-rules so that it may fix nested media
@@ -24,9 +26,12 @@ export default [
 	'gradients-interpolation-method', // run before all color functions
 	'color-mix-variadic-function-arguments', // run before any other color functions, including base color-mix
 	'color-mix', // run before any other color functions
+	'contrast-color-function',
+	'alpha-function',
 	'relative-color-syntax',
 	'lab-function',
 	'oklab-function',
+	'color-function-display-p3-linear',
 	'color-function',
 	'hwb-function',
 	'color-functional-notation',
@@ -44,14 +49,18 @@ export default [
 	'overflow-property',
 	'overflow-wrap-property',
 	'place-properties',
+	'position-area-property',
 	'system-ui-font-family',
 	'font-format-keywords',
+	'font-width-property',
 	'display-two-values',
 	'content-alt-text',
 	'ic-unit',
 	'opacity-percentage',
 	'text-decoration-shorthand',
 	'unset-value',
+	'property-rule-prelude-list',
+	'syntax-descriptor-syntax-production',
 
 	// Math functions.
 	'random-function',