+++ plugins/postcss-nesting/src/editions/2021/lib/nest-rule-within-rule.ts
@@ -7,12 +7,12 @@ import type { options } from './options.js';
 import cleanupParent from '../../shared/lib/cleanup-parent.js';
 
 export default function transformNestRuleWithinRule(node: AtRule, parent: Rule, result: Result, walk: walkFunc, opts: options): void {
-	let selectors: Array<string> = [];
+	let selectors: Array<string>;
 
 	try {
 		selectors = mergeSelectors(parent.selectors, comma(node.params), opts);
 	} catch (err) {
-		node.warn(result, `Failed to parse selectors : "${parent.selector}" / "${node.params}" with message: "${(err instanceof Error) ? err.message: err}"`);
+		node.warn(result, `Failed to parse selectors : "${parent.selector}" / "${node.params}" with message: "${(err instanceof Error) ? err.message : err}"`);
 		return;
 	}