+++ packages/css-calc/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changes to CSS Calc
 
+### 3.1.1
+
+_February 13, 2026_
+
+- Fix missing whitespace between components after solving calc expressions (e.g. `calc(10px)calc(20px)` now serializes as `10px 20px`)
+
+### 3.1.0
+
+_February 12, 2026_
+
+- Add `onParseError` option to `calc()`.
+
+### 3.0.1
+
+_February 10, 2026_
+
+- Fix source positions for generated functions and comma's. These now adopt the source position of surrounding nodes.
+
+### 3.0.0
+
+_January 14, 2026_
+
+- Updated: Support for Node `20.19.0` or later (major).
+- Removed: `commonjs` API. In supported Node versions `require(esm)` will work without needing to make code changes.
+- Updated [`@csstools/css-tokenizer`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-tokenizer) to [`4.0.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-tokenizer/CHANGELOG.md#400) (major)
+- Updated [`@csstools/css-parser-algorithms`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms) to [`4.0.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms/CHANGELOG.md#400) (major)
+
 ### 2.1.4
 
 _May 27, 2025_