+++ packages/postcss-tape/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changes to PostCSS Tape
 
-### Unreleased (patch)
+### 7.0.0
+
+_January 14, 2026_
+
+- Updated: Support for Node `20.19.0` or later (major).
+- Removed: `commonjs` API. In supported Node versions `require(esm)` will work without needing to make code changes.
+
+### 6.0.1
+
+_August 22, 2025_
 
 - Add support for Node 24